move and refactor profiler
