+++ src/utils/index.ts
@@ -77,3 +77,4 @@ export * from "./CLIUtils";
 export * from "./BNUtils";
 export * from "./CCTPUtils";
 export * from "./RetryUtils";
+export * from "./Profiler";
